# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Student ERP System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admissions:
    get:
      operationId: listAdmissions
      tags: [admissions]
      summary: List all admission applications
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: List of admissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Admission"
    post:
      operationId: createAdmission
      tags: [admissions]
      summary: Submit a new admission application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdmissionBody"
      responses:
        "201":
          description: Admission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admission"
  /admissions/{id}:
    get:
      operationId: getAdmission
      tags: [admissions]
      summary: Get a single admission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Admission details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admission"
    patch:
      operationId: updateAdmissionStatus
      tags: [admissions]
      summary: Approve or reject an admission (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdmissionStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admission"
...
  /dashboard/student/{studentId}:
    get:
      operationId: getStudentDashboard
      tags: [dashboard]
      summary: Get student dashboard data
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDashboard"
  /dashboard/faculty/{facultyId}:
    get:
      operationId: getFacultyDashboard
      tags: [dashboard]
      summary: Get faculty dashboard data
      parameters:
        - name: facultyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Faculty dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacultyDashboard"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserSession"
        token:
          type: string
      required: [user]
    UserSession:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin, faculty]
        studentId:
          type: integer
          nullable: true
        facultyId:
          type: integer
          nullable: true
      required: [id, name, email, role]
    Admission:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        address:
          type: string
        course:
          type: string
        department:
          type: string
        previousSchool:
          type: string
        previousMarks:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        remarks:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, email, phone, dateOfBirth, gender, address, course, department, previousSchool, previousMarks, status, createdAt, updatedAt]
    CreateAdmissionBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        address:
          type: string
        course:
          type: string
        department:
          type: string
        previousSchool:
          type: string
        previousMarks:
          type: number
      required: [name, email, phone, dateOfBirth, gender, address, course, department, previousSchool, previousMarks]
    UpdateAdmissionStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
        remarks:
          type: string
      required: [status]
    Student:
      type: object
      properties:
        id:
          type: integer
        studentUid:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        address:
          type: string
        course:
          type: string
        department:
          type: string
        semester:
          type: string
        rollNumber:
          type: string
        enrollmentYear:
          type: integer
        admissionId:
          type: integer
          nullable: true
        createdAt:
          type: string
      required: [id, studentUid, name, email, phone, dateOfBirth, gender, address, course, department, semester, rollNumber, enrollmentYear, createdAt]
...
    FeeRecord:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        studentUid:
          type: string
        feeType:
          type: string
          enum: [tuition, hostel, exam, library, other]
        amount:
          type: number
        dueDate:
          type: string
        paidDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [paid, pending, overdue]
        receiptNumber:
          type: string
          nullable: true
        semester:
          type: string
        academicYear:
          type: string
        createdAt:
          type: string
      required: [id, studentId, studentName, studentUid, feeType, amount, dueDate, status, semester, academicYear, createdAt]
...
    HostelAllocation:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        studentUid:
          type: string
        roomId:
          type: integer
        roomNumber:
          type: string
        block:
          type: string
        floor:
          type: integer
        allocatedDate:
          type: string
        vacatedDate:
          type: string
          nullable: true
        active:
          type: boolean
      required: [id, studentId, studentName, studentUid, roomId, roomNumber, block, floor, allocatedDate, active]
...
    ExamResult:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        studentUid:
          type: string
        subject:
          type: string
        subjectCode:
          type: string
        semester:
          type: string
        academicYear:
          type: string
        internalMarks:
          type: number
        externalMarks:
          type: number
        totalMarks:
          type: number
        maxMarks:
          type: number
        grade:
          type: string
        result:
          type: string
          enum: [pass, fail, absent]
        facultyId:
          type: integer
          nullable: true
        createdAt:
          type: string
      required: [id, studentId, studentName, studentUid, subject, subjectCode, semester, academicYear, internalMarks, externalMarks, totalMarks, maxMarks, grade, result, createdAt]
...
    Faculty:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        department:
          type: string
        designation:
          type: string
        subjects:
          type: array
          items:
            type: string
        joinedAt:
          type: string
      required: [id, name, email, phone, department, designation, subjects, joinedAt]
...
    StudentDashboard:
      type: object
      properties:
        student:
          $ref: "#/components/schemas/Student"
        admission:
          $ref: "#/components/schemas/Admission"
          nullable: true
        fees:
          type: array
          items:
            $ref: "#/components/schemas/FeeRecord"
        hostelAllocation:
          $ref: "#/components/schemas/HostelAllocation"
          nullable: true
        examResults:
          type: array
          items:
            $ref: "#/components/schemas/ExamResult"
        cgpa:
          type: number
          nullable: true
      required: [student, fees, examResults]
    FacultyDashboard:
      type: object
      properties:
        faculty:
          $ref: "#/components/schemas/Faculty"
        totalStudentsManaged:
          type: integer
        recentResults:
          type: array
          items:
            $ref: "#/components/schemas/ExamResult"
        subjectPassRates:
          type: array
          items:
            type: object
            properties:
              subject:
                type: string
              passRate:
                type: number
              totalStudents:
                type: integer
            required: [subject, passRate, totalStudents]
      required: [faculty, totalStudentsManaged, recentResults, subjectPassRates]