# lib/api-zod/src/generated/types/facultyDashboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student ERP System API
 * OpenAPI spec version: 0.1.0
 */
import type { ExamResult } from "./examResult";
import type { Faculty } from "./faculty";
import type { FacultyDashboardSubjectPassRatesItem } from "./facultyDashboardSubjectPassRatesItem";

export interface FacultyDashboard {
  faculty: Faculty;
  totalStudentsManaged: number;
  recentResults: ExamResult[];
  subjectPassRates: FacultyDashboardSubjectPassRatesItem[];
}