# lib/api-zod/src/generated/types/feeRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student ERP System API
 * OpenAPI spec version: 0.1.0
 */
import type { FeeRecordFeeType } from "./feeRecordFeeType";
import type { FeeRecordStatus } from "./feeRecordStatus";

export interface FeeRecord {
  id: number;
  studentId: number;
  studentName: string;
  studentUid: string;
  feeType: FeeRecordFeeType;
  amount: number;
  dueDate: string;
  paidDate?: string | null;
  status: FeeRecordStatus;
  receiptNumber?: string | null;
  semester: string;
  academicYear: string;
  createdAt: string;
}