# lib/api-zod/src/generated/types/hostelAllocation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student ERP System API
 * OpenAPI spec version: 0.1.0
 */

export interface HostelAllocation {
  id: number;
  studentId: number;
  studentName: string;
  studentUid: string;
  roomId: number;
  roomNumber: string;
  block: string;
  floor: number;
  allocatedDate: string;
  vacatedDate?: string | null;
  active: boolean;
}