# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student ERP System API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  studentUid: string;
  name: string;
  email: string;
  phone: string;
  dateOfBirth: string;
  gender: string;
  address: string;
  course: string;
  department: string;
  semester: string;
  rollNumber: string;
  enrollmentYear: number;
  admissionId?: number | null;
  createdAt: string;
}