# lib/api-zod/src/generated/types/studentDashboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student ERP System API
 * OpenAPI spec version: 0.1.0
 */
import type { Admission } from "./admission";
import type { ExamResult } from "./examResult";
import type { FeeRecord } from "./feeRecord";
import type { HostelAllocation } from "./hostelAllocation";
import type { Student } from "./student";

export interface StudentDashboard {
  student: Student;
  admission?: Admission | null;
  fees: FeeRecord[];
  hostelAllocation?: HostelAllocation | null;
  examResults: ExamResult[];
  cgpa?: number | null;
}